package/opensbi: add ability to build using external platform definition.
OpenSBI supports building against a platform definition which is out of
tree by setting the PLATFORM_DIR environment variable. When porting
builroot to a new board this is simpler than patching in a platform
directory to the OpenSBI source, or building a tarball.

This patch brings the abilty to use an external platform definition
when building OpenSBI through buildroot.